+++ changelog.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Removed** unused `app-root.css` file and reference
 - **Fixed** [#49](https://github.com/codyburleson/cburleson-pwa/issues/49) Add page: access-the-was-console-from-websphere-portal-server
 - **Fixed** [#45](https://github.com/codyburleson/cburleson-pwa/issues/45) Add page: resolving-404-errors-accessing-wcm-content-through-poc-servlet
-## [1.1.1] - 2019-12-25
+- **Fixed** [#44](https://github.com/codyburleson/cburleson-pwa/issues/44) Add page: notes-on-zbrush-slicecurve-brush/
 
-### Added
+## [1.1.1] - 2019-12-25
 
 - **Added** `robots.txt` file allowing all agents to crawl all resources, and pointing to the `sitemap.txt` file.
 - **Added** 3 `<link rel="preconnect" href="..."/>` tags based on performance improvement suggestion from Lighthouse.